donate
donate works. does not check if same community. also made minor changes
in create and apparently search.

IMPORTANT: in the database, i created a new field for the projects
table: currentFunds. currentFunds is of type decimal with precision 10
and scale 2, with default value of 0. I also changed goal to decimal
with precision 10 and scale 2.